Bump fusion api to 1.21.2
This commit removes the example modules in favor of me writing unit tests instead as that would likely be the better notion, it simplifies the build script and project structure at the same time.

I've extracted the addons code into a separate module as it should be standalone since it is a generic class loader that doesn't rely on Minecraft code.

The ComponentBuilder was removed as it seemed unnecessary for the use case I had it for, I'll just find other ways.

The common module will always have Adventure API, and when I inevitably expand to other platforms like Discord, I'll just shade in Adventure API as needed and call it a day. It's platform-agnostic anyway.